Seperated tests in ExtendedNextflowIT.java into two files
