+++ dockstore-integration-testing/src/test/java/io/dockstore/client/cli/ExtendedNextflowIT.java
@@ -17,14 +17,12 @@ package io.dockstore.client.cli;
 
 import static org.junit.Assert.assertNotSame;
 
-import io.dockstore.common.BitBucketTest; // To delete
 import io.dockstore.common.CommonTestUtilities;
 import io.dockstore.common.ConfidentialTest;
 import io.dockstore.common.DescriptorLanguage;
 import io.dockstore.common.SourceControl;
 import io.dockstore.common.WorkflowTest;
 import io.dockstore.webservice.DockstoreWebserviceApplication;
-import io.dockstore.webservice.core.SourceFile;
 import io.dockstore.webservice.jdbi.FileDAO;
 import io.swagger.client.ApiClient;
 import io.swagger.client.api.UsersApi;
@@ -43,6 +41,7 @@ import org.junit.contrib.java.lang.system.ExpectedSystemExit;
 import org.junit.experimental.categories.Category;
 import org.junit.rules.ExpectedException;
 
+@Category({ ConfidentialTest.class, WorkflowTest.class })
 public class ExtendedNextflowIT extends BaseIT {
 
     // workflow with a bin directory
@@ -74,8 +73,7 @@ public class ExtendedNextflowIT extends BaseIT {
         ManagedSessionContext.bind(session);
     }
 
-    @Test 
-    @Category({ ConfidentialTest.class, WorkflowTest.class })
+    @Test
     public void testNextflowSecondaryFiles() throws Exception {
         CommonTestUtilities.cleanStatePrivate1(SUPPORT);
         final ApiClient webClient = getWebClient(USER_1_USERNAME, testingPostgres);
@@ -110,105 +108,9 @@ public class ExtendedNextflowIT extends BaseIT {
         Assert.assertEquals("Fast automated prediction of protein antimicrobial regions", refreshGithub.getDescription());
     }
 
-    @Test 
-    @Category(BitBucketTest.class)
-    public void testBitbucketNextflowWorkflow() throws Exception {
-        CommonTestUtilities.cleanStatePrivate2(SUPPORT, false);
-        final ApiClient webClient = getWebClient(USER_2_USERNAME, testingPostgres);
-        WorkflowsApi workflowApi = new WorkflowsApi(webClient);
-        UsersApi usersApi = new UsersApi(webClient);
-        User user = usersApi.getUser();
-        // get workflow stubs
-        Workflow workflow = workflowApi.manualRegister(SourceControl.BITBUCKET.name(), "dockstore_testuser2/ampa-nf", "/nextflow.config", "",
-                DescriptorLanguage.NEXTFLOW.getShortName(), "/foo.json");
-        workflowApi.refresh(workflow.getId(), false);
-
-        // do targeted refresh, should promote workflow to fully-fleshed out workflow
-        Workflow workflowByPathBitbucket = workflowApi.getWorkflowByPath(DOCKSTORE_TEST_USER_NEXTFLOW_BITBUCKET_WORKFLOW, BIOWORKFLOW, "versions");
-        // need to set paths properly
-        workflowByPathBitbucket.setWorkflowPath("/nextflow.config");
-        workflowByPathBitbucket.setDescriptorType(Workflow.DescriptorTypeEnum.NFL);
-        workflowApi.updateWorkflow(workflowByPathBitbucket.getId(), workflowByPathBitbucket);
-        workflowByPathBitbucket = workflowApi.getWorkflowByPath(DOCKSTORE_TEST_USER_NEXTFLOW_BITBUCKET_WORKFLOW, BIOWORKFLOW, "versions");
-        final Workflow bitbucketWorkflow = workflowApi.refresh(workflowByPathBitbucket.getId(), false);
-        Workflow byPathWorkflow = workflowApi.getWorkflowByPath(DOCKSTORE_TEST_USER_NEXTFLOW_BITBUCKET_WORKFLOW, BIOWORKFLOW, "versions");
-        // There are 3 versions: master, v1.0, and v2.0
-        // master and v2.0 has a nextflow.config file that has description and author, v1.0 does not
-        // v1.0 will pull description from README instead but the others will use nextflow.config. v1.0 also does not have a valid descriptor (or any sourcefiles)
-        testWorkflowVersionMetadata(bitbucketWorkflow);
-        testWorkflowVersionMetadata(byPathWorkflow);
-        // Purposely mess up the metadata to test if it can be updated through refresh
-        testingPostgres.runUpdateStatement("update version_metadata set email='bad_potato'");
-        testingPostgres.runUpdateStatement("update version_metadata set author='bad_potato'");
-        testingPostgres.runUpdateStatement("update version_metadata set description='bad_potato'");
-        final Workflow refreshedBitbucketWorkflow = workflowApi.refresh(workflowByPathBitbucket.getId(), true);
-        byPathWorkflow = workflowApi.getWorkflowByPath(DOCKSTORE_TEST_USER_NEXTFLOW_BITBUCKET_WORKFLOW, BIOWORKFLOW, "versions");
-        // This tests if it can fix outdated metadata
-        testWorkflowVersionMetadata(refreshedBitbucketWorkflow);
-        testWorkflowVersionMetadata(byPathWorkflow);
-        List<io.dockstore.webservice.core.SourceFile> sourceFileList = fileDAO.findSourceFilesByVersion(bitbucketWorkflow.getWorkflowVersions().stream().filter(version -> version.getName().equals("v2.0")).findFirst().get().getId());
-        Assert.assertEquals(4, sourceFileList.size());
-    }
-
-    /**
-     * This tests the DOCKSTORE_TEST_USER_NEXTFLOW_BITBUCKET_WORKFLOW metadata is correct after a refresh
-     * @param workflow  The DOCKSTORE_TEST_USER_NEXTFLOW_BITBUCKET_WORKFLOW workflow
-     */
-    private void testWorkflowVersionMetadata(Workflow workflow) {
-
-        final String partialReadmeDescription = "AMPA-NF is a pipeline for assessing the antimicrobial domains of proteins,";
-        final String descriptorDescription = "Fast automated prediction of protein antimicrobial regions";
-        final String versionWithReadmeDescription = "v1.0";
-
-        Assert.assertEquals(descriptorDescription, workflow.getDescription());
-        Assert.assertTrue(workflow.getWorkflowVersions().stream().anyMatch(workflowVersion -> workflowVersion.getName().equals(versionWithReadmeDescription)));
-        workflow.getWorkflowVersions().forEach(workflowVersion -> {
-            if (workflowVersion.getName().equals(versionWithReadmeDescription)) {
-                Assert.assertTrue(workflowVersion.getDescription().contains(partialReadmeDescription));
-                Assert.assertNull(workflowVersion.getAuthor());
-                Assert.assertNull(workflowVersion.getEmail());
-            } else {
-                Assert.assertNotNull(descriptorDescription, workflowVersion.getDescription());
-                Assert.assertEquals("test.user@test.com", workflowVersion.getAuthor());
-                Assert.assertNull(workflowVersion.getEmail());
-            }
-        });
-    }
-
-    @Test 
-    @Category({ ConfidentialTest.class, WorkflowTest.class })
+    @Test
     public void testGitlabNextflowWorkflow() {
         // TODO: need to look into the SlowTest situation but we also need to reactivate the tests against API V4 for 1.5.0
-    }
-
-    @Test 
-    @Category(BitBucketTest.class)
-    public void testBitbucketBinaryWorkflow() throws Exception {
-        CommonTestUtilities.cleanStatePrivate2(SUPPORT, false);
-        final ApiClient webClient = getWebClient(USER_2_USERNAME, testingPostgres);
-        WorkflowsApi workflowApi = new WorkflowsApi(webClient);
-        // get workflow stubs
-
-        Workflow workflow = workflowApi
-                .manualRegister(SourceControl.BITBUCKET.name(), "dockstore_testuser2/kallisto-nf", "/nextflow.config", "",
-                        DescriptorLanguage.NEXTFLOW.getShortName(), "/foo.json");
-        workflowApi.refresh(workflow.getId(), false);
-
-        // do targeted refresh, should promote workflow to fully-fleshed out workflow
-        Workflow workflowByPathGithub = workflowApi.getWorkflowByPath(DOCKSTORE_TEST_USER_NEXTFLOW_BINARY_WORKFLOW, BIOWORKFLOW, null);
-        // need to set paths properly
-        workflowByPathGithub.setWorkflowPath("/nextflow.config");
-        workflowByPathGithub.setDescriptorType(Workflow.DescriptorTypeEnum.NFL);
-        workflowApi.updateWorkflow(workflowByPathGithub.getId(), workflowByPathGithub);
-
-        workflowByPathGithub = workflowApi.getWorkflowByPath(DOCKSTORE_TEST_USER_NEXTFLOW_BINARY_WORKFLOW, BIOWORKFLOW, null);
-        final Workflow bitbucketWorkflow = workflowApi.refresh(workflowByPathGithub.getId(), false);
-        Assert.assertTrue("Should have gotten the description from README", bitbucketWorkflow.getDescription().contains("A Nextflow implementation of Kallisto & Sleuth RNA-Seq Tools"));
-        List<SourceFile> sourceFileList = fileDAO.findSourceFilesByVersion(bitbucketWorkflow.getWorkflowVersions().stream().filter(version -> version.getName().equals("v1.0")).findFirst().get().getId());
-        Assert.assertEquals(6, sourceFileList.size());
-        // two of the files should essentially be blanked
-        Assert.assertEquals("two files have our one-line warning", 2, sourceFileList.stream()
-            .filter(file -> file.getContent().split("\n").length == 1 && file.getContent().contains("Dockstore does not")).count());
     }
 
 }
